Add versioning to airdrop, nfts and yield farming scripts
