+++ scripts/6_deployAirdrop.js
@@ -12,34 +12,47 @@ const { parseBalanceMap } = require("../helpers/parse-balance-map")
 
 const { AIRDROP_INFO } = require("./constants")
 
+const { bn, tokens, bnToInt, timeInDays, timeInDate, readArgumentsFile, deployContract, verifyAllContracts } = require('../helpers/utils')
+const { versionVerifier, upgradeVersion } = require('../loaders/versionVerifier')
+
 // Primary Function
 async function main() {
+  // Version Check
+  console.log(chalk.bgBlack.bold.green(`\n✌️  Running Version Checks \n-----------------------\n`))
+  const versionDetails = versionVerifier(["pushTokenAddress"])
+  console.log(chalk.bgWhite.bold.black(`\n\t\t\t\n Version Control Passed \n\t\t\t\n`))
+
   // First deploy all contracts
   console.log(chalk.bgBlack.bold.green(`\n📡 Deploying MerkleDistributor \n-----------------------\n`));
-  const deployedContracts = await setupAllContracts();
+  const deployedContracts = await setupAllContracts(versionDetails);
   console.log(chalk.bgWhite.bold.black(`\n\t\t\t\n All Contracts Deployed \n\t\t\t\n`));
 
   // Try to verify
   console.log(chalk.bgBlack.bold.green(`\n📡 Verifying Contracts \n-----------------------\n`));
-  await verifyAllContracts(deployedContracts);
+  await verifyAllContracts(deployedContracts, versionDetails);
   console.log(chalk.bgWhite.bold.black(`\n\t\t\t\n All Contracts Verified \n\t\t\t\n`));
+
+  // Upgrade Version
+  console.log(chalk.bgBlack.bold.green(`\n📟 Upgrading Version   \n-----------------------\n`))
+  upgradeVersion()
+  console.log(chalk.bgWhite.bold.black(`\n\t\t\t\n ✅ Version upgraded    \n\t\t\t\n`))
 }
 
 // Secondary Functions
 // Deploy All Contracts
-async function setupAllContracts() {
+async function setupAllContracts(versionDetails) {
   let deployedContracts = [];
 
   // Generate Merkle Root
   const merkleRoot = await generateMerkleRoot()
 
   // Deploy MerkleDistributor
-  const DistributorArgs = [AIRDROP_INFO.airdrop.pushTokenAddress, merkleRoot]
+  const DistributorArgs = [versionDetails.deploy.args.pushTokenAddress, merkleRoot]
   const Distributor = await deployContract("MerkleDistributor", DistributorArgs, "MerkleDistributor")
   deployedContracts.push(Distributor)
 
   // Push token transfer to MerkleDistributor
-  await tokensToDistrbutor(Distributor, AIRDROP_INFO.airdrop.pushTokenAddress)
+  await tokensToDistrbutor(Distributor, versionDetails.deploy.args.pushTokenAddress)
 
   return deployedContracts;
 }
@@ -71,71 +84,6 @@ async function tokensToDistrbutor(Distributor, pushTokenAddress) {
   console.log(chalk.bgBlack.white(`Transaction hash:`), chalk.gray(`${tx.hash}`))
   console.log(chalk.bgBlack.white(`Transaction etherscan:`), chalk.gray(`https://${hre.network.name}.etherscan.io/tx/${tx.hash}`))
 
-}
-
-// Verify All Contracts
-async function verifyAllContracts(deployedContracts) {
-  return new Promise(async function(resolve, reject) {
-
-    for await (contract of deployedContracts) {
-      const arguments = contract.deployargs
-
-      if (hre.network.name != "hardhat") {
-        // Mostly a real network, verify
-        const { spawnSync } = require( 'child_process' )
-        const ls = spawnSync( `npx`, [ 'hardhat', 'verify', '--network', hre.network.name, contract.address ] )
-
-        console.log( `stderr: ${ ls.stderr.toString() }` )
-        console.log( `stdout: ${ ls.stdout.toString() }` )
-      }
-      else {
-        console.log(chalk.bgWhiteBright.black(`${contract.filename}.sol`), chalk.bgRed.white(` is on Hardhat network... skipping`))
-      }
-    }
-
-    resolve();
-  });
-}
-
-// Helper Functions
-// For Deploy
-async function deploy(name, _args, identifier) {
-  const args = _args || [];
-
-  console.log(`📄 ${name}`);
-  const contractArtifacts = await ethers.getContractFactory(name);
-  const contract = await contractArtifacts.deploy(...args);
-  await contract.deployed()
-  console.log(
-    chalk.cyan(name),
-    "deployed to:",
-    chalk.magenta(contract.address)
-  );
-  fs.writeFileSync(`artifacts/${name}_${identifier}.address`, contract.address);
-  return contract;
-}
-
-async function deployContract(contractName, contractArgs, identifier) {
-  let contract = await deploy(contractName, contractArgs, identifier);
-
-  contract.filename = `${contractName} -> ${identifier}`;
-  contract.deployargs = contractArgs;
-
-  return contract;
-}
-
-function readArgumentsFile(contractName) {
-  let args = [];
-  try {
-    const argsFile = `./contracts/${contractName}.args`
-    if (fs.existsSync(argsFile)) {
-      args = JSON.parse(fs.readFileSync(argsFile));
-    }
-  } catch (e) {
-    console.log(e);
-  }
-
-  return args;
 }
 
 // We recommend this pattern to be able to use async/await everywhere
