+++ scripts/7_deployNFTs.js
@@ -9,26 +9,39 @@ const fs = require("fs");
 const chalk = require("chalk");
 const { config, ethers } = require("hardhat");
 
+const { bn, tokens, bnToInt, timeInDays, timeInDate, deployContract, verifyAllContracts, distributeInitialFunds } = require('../helpers/utils')
+const { versionVerifier, upgradeVersion } = require('../loaders/versionVerifier')
+
 const {
   NFT_INFO
 } = require("./constants")
 
 // Primary Function
 async function main() {
+  // Version Check
+  console.log(chalk.bgBlack.bold.green(`\n✌️  Running Version Checks \n-----------------------\n`))
+  const versionDetails = versionVerifier(["pushTokenAddress"])
+  console.log(chalk.bgWhite.bold.black(`\n\t\t\t\n Version Control Passed \n\t\t\t\n`))
+
   // First deploy all contracts
   console.log(chalk.bgBlack.bold.green(`\n📡 Deploying ROCKSTAR NFTs and Minting \n-----------------------\n`));
-  const deployedContracts = await setupAllContracts();
+  const deployedContracts = await setupAllContracts(versionDetails)
   console.log(chalk.bgWhite.bold.black(`\n\t\t\t\n All Contracts Deployed \n\t\t\t\n`));
 
   // Try to verify
   console.log(chalk.bgBlack.bold.green(`\n📡 Verifying Contracts \n-----------------------\n`));
-  await verifyAllContracts(deployedContracts);
+  await verifyAllContracts(deployedContracts, versionDetails)
   console.log(chalk.bgWhite.bold.black(`\n\t\t\t\n All Contracts Verified \n\t\t\t\n`));
+
+  // Upgrade Version
+  console.log(chalk.bgBlack.bold.green(`\n📟 Upgrading Version   \n-----------------------\n`))
+  upgradeVersion()
+  console.log(chalk.bgWhite.bold.black(`\n\t\t\t\n ✅ Version upgraded    \n\t\t\t\n`))
 }
 
 // Secondary Functions
 // Deploy All Contracts
-async function setupAllContracts() {
+async function setupAllContracts(versionDetails) {
   let deployedContracts = [];
   const signer = await ethers.getSigner(0)
 
@@ -41,7 +54,7 @@ async function setupAllContracts() {
   deployedContracts.push(BatchMintNFT)
 
   // Deploy NFTRewards
-  const NFTRewardsArgs = [bn(NFT_INFO.nfts.tokens).div(bn(NFT_INFO.nfts.users)), NFT_INFO.nfts.pushTokenAddress, Rockstar.address]
+  const NFTRewardsArgs = [bn(NFT_INFO.nfts.tokens).div(bn(NFT_INFO.nfts.users)), versionDetails.deploy.args.pushTokenAddress, Rockstar.address]
   const NFTRewards = await deployContract("NFTRewards", NFTRewardsArgs, "RockstarNFTRewards")
   deployedContracts.push(NFTRewards)
 
@@ -49,7 +62,7 @@ async function setupAllContracts() {
   await batchMintNFTs(Rockstar, BatchMintNFT)
 
   // Push token transfer to NFTRewards
-  await tokensToNFTRewards(NFTRewards, NFT_INFO.nfts.pushTokenAddress)
+  await tokensToNFTRewards(NFTRewards, versionDetails.deploy.args.pushTokenAddress)
 
   // return deployed contracts
   return deployedContracts;
@@ -111,71 +124,6 @@ async function batchMintNFTs(rockstar, batchMintNFT) {
 
   console.log(chalk.bgBlack.white(`Transaction hash:`), chalk.gray(`${tx.hash}`))
   console.log(chalk.bgBlack.white(`Transaction etherscan:`), chalk.gray(`https://${hre.network.name}.etherscan.io/tx/${tx.hash}`))
-}
-
-// Verify All Contracts
-async function verifyAllContracts(deployedContracts) {
-  return new Promise(async function(resolve, reject) {
-
-    for await (contract of deployedContracts) {
-      const arguments = contract.deployargs
-
-      if (hre.network.name != "hardhat") {
-        // Mostly a real network, verify
-        const { spawnSync } = require( 'child_process' )
-        const ls = spawnSync( `npx`, [ 'hardhat', 'verify', '--network', hre.network.name, contract.address ] )
-
-        console.log( `stderr: ${ ls.stderr.toString() }` )
-        console.log( `stdout: ${ ls.stdout.toString() }` )
-      }
-      else {
-        console.log(chalk.bgWhiteBright.black(`${contract.filename}.sol`), chalk.bgRed.white(` is on Hardhat network... skipping`))
-      }
-    }
-
-    resolve();
-  });
-}
-
-// Helper Functions
-// For Deploy
-async function deploy(name, _args, identifier) {
-  const args = _args || [];
-
-  console.log(`📄 ${name}`);
-  const contractArtifacts = await ethers.getContractFactory(name);
-  const contract = await contractArtifacts.deploy(...args);
-  await contract.deployed()
-  console.log(
-    chalk.cyan(name),
-    "deployed to:",
-    chalk.magenta(contract.address)
-  );
-  fs.writeFileSync(`artifacts/${name}_${identifier}.address`, contract.address);
-  return contract;
-}
-
-async function deployContract(contractName, contractArgs, identifier) {
-  let contract = await deploy(contractName, contractArgs, identifier);
-
-  contract.filename = `${contractName} -> ${identifier}`;
-  contract.deployargs = contractArgs;
-
-  return contract;
-}
-
-function readArgumentsFile(contractName) {
-  let args = [];
-  try {
-    const argsFile = `./contracts/${contractName}.args`
-    if (fs.existsSync(argsFile)) {
-      args = JSON.parse(fs.readFileSync(argsFile));
-    }
-  } catch (e) {
-    console.log(e);
-  }
-
-  return args;
 }
 
 // We recommend this pattern to be able to use async/await everywhere
