+++ scripts/versioncontrol/6_deployAirdrop.config.js
@@ -1,0 +1,30 @@
+const deploy = {
+  network: {
+    mainnet: {
+      version: 2
+    },
+    goerli: {
+      version: 2
+    },
+    kovan: {
+      version: 2
+    },
+    ropsten: {
+      version: 2
+    },
+    rinkeby: {
+      version: 2
+    },
+    hardhat: {
+      version: 2
+    },
+    localhost: {
+      version: 2
+    }
+  },
+  args: {
+    pushTokenAddress: null
+  }
+}
+
+exports.deploy = deploy
