+++ scripts/versioncontrol/7_deployNFTs.config.js
@@ -1,0 +1,30 @@
+const deploy = {
+  network: {
+    mainnet: {
+      version: 1
+    },
+    goerli: {
+      version: 1
+    },
+    kovan: {
+      version: 1
+    },
+    ropsten: {
+      version: 1
+    },
+    rinkeby: {
+      version: 1
+    },
+    hardhat: {
+      version: 1
+    },
+    localhost: {
+      version: 1
+    }
+  },
+  args: {
+    pushTokenAddress: null
+  }
+}
+
+exports.deploy = deploy
