+++ scripts/versioncontrol/8_deployYieldFarm.config.js
@@ -1,0 +1,32 @@
+const deploy = {
+  network: {
+    mainnet: {
+      version: 1,
+    },
+    goerli: {
+      version: 1,
+    },
+    kovan: {
+      version: 1,
+    },
+    ropsten: {
+      version: 1,
+    },
+    rinkeby: {
+      version: 1,
+    },
+    hardhat: {
+      version: 1,
+    },
+    localhost: {
+      version: 1,
+    }
+  },
+  args: {
+    pushTokenAddress: '0x38D542f47b1B949146e3961eCd87872fdEA49679',
+    communityVaultAddress: '0x29d55CF846c094d2899Dcd596E84874271C46f88',
+    lpPoolTokenAddress: '0x38D542f47b1B949146e3961eCd87872fdEA49679',
+  }
+}
+
+exports.deploy = deploy
